move signal connection to end of models #75
