+++ loc_insurancemaps/apps.py
@@ -14,4 +14,3 @@ class AppConfig(BaseAppConfig):
     def ready(self):
         super(AppConfig, self).ready()
         run_setup_hooks()
-        import loc_insurancemaps.signals
